+++ HelpMeCook/app/src/main/java/br/com/helpmecook/sqlite/CookbookDAO.java
@@ -43,7 +43,7 @@ public class CookbookDAO {
         return database.insert(TABLE_NAME, null, values);
     }
 
-    public boolean delete(int id) {
+    public boolean delete(long id) {
         long wReturn = database.delete(TABLE_NAME, ID + " = " + id, null);
 
         if (wReturn == 0) {
@@ -53,9 +53,9 @@ public class CookbookDAO {
         }
     }
 
-    public List<Integer> readAll() {
-        int id;
-        ArrayList<Integer> recipes = new ArrayList<Integer>();
+    public List<Long> readAll() {
+        long id;
+        ArrayList<Long> recipes = new ArrayList<Long>();
 
         Cursor c = database.query(TABLE_NAME, allColumns, null, null, null, null, null);
 
@@ -63,7 +63,7 @@ public class CookbookDAO {
             int indexId = c.getColumnIndex(ID);
 
             do {
-                id = c.getInt(indexId);
+                id = c.getLong(indexId);
                 recipes.add(id);
             } while(c.moveToNext());
         }
